refactor: data refactor; the data.ts file was removed and the data was broken down into several smaller files adequate for the various components of the site; the data structure was prepared for additional trails whose data we do not yet have access to;
