+++ src/app/szlak-partyzancki/data/partners.ts
@@ -1,0 +1,13 @@
+export const partners: string[] = [
+  "Powiat Dębicki",
+  "Miasto Dębica",
+  "Gmina Dębica, Gmina Pilzno",
+  "Miejski Zarząd Oświaty w Dębicy",
+  "Centrum Kultury i Bibliotek Gminy Dębica",
+  "Lasy Państwowe – Nadleśnictwo Dębica",
+  "Katolickie Centrum Edukacji „KANA” w Dębicy",
+  "Miejska Komunikacja Samochodowa Sp. z o.o. w Dębicy",
+  "Hufiec Dębica Związku Harcerstwa Polskiego im. Dębickich Szarych Szeregów",
+  "Hufiec Pilzno Związku Harcerstwa Polskiego im. Aleksandra Kamińskiego ps. „Kamyk”",
+  "Związek Strzelecki „Strzelec” JS 2017",
+] as const;
